implemented new node creation and applied styling for it

+++ app/org/org-node-detail/org-node-detail.component.ts
@@ -13,8 +13,10 @@ import { OrgNodeModel, OrgService } from "../shared/index";
 
 export class OrgNodeDetailComponent {
     @Input() selectedOrgNode: OrgNodeModel;
+    @Input() isAddMode: boolean;
     @Output() deleteNode = new EventEmitter<OrgNodeModel>();
     @Output() updateNode = new EventEmitter<OrgNodeModel>();
+    @Output() addNode = new EventEmitter<OrgNodeModel>();
 
     private isEditMode: boolean;
     private editNodeDetails: OrgNodeModel;
@@ -30,17 +32,51 @@ export class OrgNodeDetailComponent {
         return (node.children != null);
     }
 
+    private isNullOREmpty(value: string) {
+        if (value && value.trim().length > 0) {
+            return false;
+        }
+        return true;
+    }
+
     private onSubmit(form: NgForm) {
         let data = JSON.stringify(form.value, null, 2);
-        this.editNodeDetails = new OrgNodeModel();
-        this.editNodeDetails.NodeFirstName = form.value.firstName;
-        this.editNodeDetails.NodeLastName = form.value.lastName;
-        this.editNodeDetails.Description = form.value.description;
-        this.editNodeDetails.children = this.selectedOrgNode.children;
-        this.editNodeDetails.NodeID = this.selectedOrgNode.NodeID;
-        this.editNodeDetails.OrgID = this.selectedOrgNode.OrgID;
-        this.editNodeDetails.ParentNodeID = this.selectedOrgNode.ParentNodeID;
-        this.editNode(this.editNodeDetails);
+
+        if (!this.isNullOREmpty(form.value.firstName)) {
+            this.editNodeDetails = new OrgNodeModel();
+            this.editNodeDetails.NodeFirstName = form.value.firstName;
+            this.editNodeDetails.NodeLastName = form.value.lastName;
+            this.editNodeDetails.Description = form.value.description;
+            this.editNodeDetails.children = this.selectedOrgNode.children;
+            this.editNodeDetails.NodeID = this.selectedOrgNode.NodeID;
+            this.editNodeDetails.OrgID = this.selectedOrgNode.OrgID;
+            this.editNodeDetails.ParentNodeID = this.selectedOrgNode.ParentNodeID;
+
+            if (this.isAddMode) {
+                this.editNodeDetails.NodeID = 310;
+                this.emitaddNodeNotification(this.editNodeDetails);
+                // this.addNewNode(this.editNodeDetails);
+            } else {
+                this.editNode(this.editNodeDetails);
+            }
+        }
+    }
+
+    private emitaddNodeNotification(data: OrgNodeModel) {
+        if (data) {
+            // console.log(data);
+            this.addNode.emit(data);
+        }
+    }
+
+    private addNewNode(node: OrgNodeModel) {
+        if (!node) { return; }
+        // we don"t really need to send any child info to the server at this point
+        node.children = null;
+        this.orgService.updateNode(node)
+            .subscribe(data => this.emitUpdateNodeNotification(data),
+            error => this.handleError(error),
+            () => console.log("Node Updated Complete"));
     }
 
     private editNode(node: OrgNodeModel) {
@@ -62,8 +98,11 @@ export class OrgNodeDetailComponent {
     }
 
     private onDeleteNodeClicked() {
-        if (this.selectedOrgNode) {
-            if (this.selectedOrgNode.children == null) {
+        if (this.isAddMode) {
+            this.deleteNode.emit(this.selectedOrgNode);
+        }
+        else {
+            if (this.selectedOrgNode.children == null && !this.isAddMode) {
                 this.orgService.deleteNode(this.selectedOrgNode.NodeID)
                     .subscribe(data => this.emitDeleteNodeNotification(data),
                     error => this.handleError(error),
